Adds VAPI voice assistant integration
Introduces an example integration with the VAPI platform,
showcasing the use of Speechmatics for STT in a voice assistant.

This includes example code, docker configuration, and documentation
to guide users in setting up and testing a voice assistant leveraging
VAPI and Speechmatics.

The example provides functionalities for assistant creation, listing,
and deletion via the VAPI Server SDK.